+++ brewtils/plugin.py
@@ -13,7 +13,8 @@ from pathlib import Path
 import appdirs
 from box import Box
 from datetime import datetime, timezone
-from packaging.version import InvalidVersion, parse, Version
+from packaging.requirements import Requirement
+from packaging.version import Version
 from requests import ConnectionError as RequestsConnectionError
 
 import brewtils
@@ -394,97 +395,29 @@ class Plugin(object):
             return current_timestamp + add_time
         return current_timestamp
 
-    def parse_require_version(self, require):
-        """
-        Follows npm version symbols
-        ^	accept updates to minor and patch releases only.	^0.13.0: 0.13.1, 0.14.0
-        ~	accept updates to patch releases only.	~0.13.0: 0.13.1 (not 0.14.0)
-        >	accept updates to any version greater than specified.	>0.13.0: 0.13.1, 0.14.1, 1.1.1
-        <	accept updates to any version less than specified.	<3.0.0: 2.0.0, 2.9.0
-        >=	accept any version greater than or equal to specified.	>=3.0.0: 3.0.0, 4.1.0
-        <=	accept any version less than or equal to specified.	<=3.0.0: 3.0.0, 2.9.0
-        =	accept only the exact specified version.	=3.0.0: 3.0.0, (not 3.0.1)
-        Future: regex version matching
-        """
-        name_version_list = re.findall(r"(\w+)([\^~]|[<>]=?|=)(.*)", require)
-        if len(name_version_list) > 1:
-            raise ValueError("Failed to parse name and version")
-        elif len(name_version_list) == 1:
-            return name_version_list[0]
-        else:
-            return require, None, None
-
     def get_system_matching_version(self, require, **kwargs):
-        """
-        Get system matching version or None.
-        Kwargs to accept named parameters like filter_latest, filter_running, and local
-        """
-        name_version = self.parse_require_version(require)
-        require_name = name_version[0]
-        require_type = name_version[1]
-        require_version = name_version[2]
+        system = None
+        req = Requirement(require)
+        require_name = req.name
+        require_version = req.specifier
         if require_version:
-            try:
-                parsed_version = str(parse(require_version))
-                if require_type == "=":
-                    system = self._ez_client.find_unique_system(
-                        name=require_name, version=parsed_version, **kwargs
-                    )
-                else:
-                    pattern = re.compile(
-                        r"^(?:([0-9]+)\.([0-9]+)\.([0-9]+)){1}(?:\.([A-Za-z0-9]+)){0,1}$"
-                    )
-                    match = re.search(pattern, parsed_version)
-                    if match:
-                        major, minor, _, _ = match.groups()
-                        if require_type == "^" and major:
-                            system = self._ez_client.find_unique_system(
-                                name=require_name,
-                                version__startswith=f"{major}.",
-                                version__gte=parsed_version,
-                                **kwargs,
-                            )
-                        elif require_type == "~" and major and minor:
-                            system = self._ez_client.find_unique_system(
-                                name=require_name,
-                                version__startswith=f"{major}.{minor}.",
-                                version__gte=parsed_version,
-                                **kwargs,
-                            )
-                        elif require_type == ">":
-                            system = self._ez_client.find_unique_system(
-                                name=require_name,
-                                version__gt=parsed_version,
-                                **kwargs,
-                            )
-                        elif require_type == ">=":
-                            system = self._ez_client.find_unique_system(
-                                name=require_name,
-                                version__gte=parsed_version,
-                                **kwargs,
-                            )
-                        elif require_type == "<":
-                            system = self._ez_client.find_unique_system(
-                                name=require_name,
-                                version__lt=parsed_version,
-                                **kwargs,
-                            )
-                        elif require_type == "<=":
-                            system = self._ez_client.find_unique_system(
-                                name=require_name,
-                                version__lte=parsed_version,
-                                **kwargs,
-                            )
-            except InvalidVersion:
-                # TODO: Regex. Switch to use version__regex when mongoengine>=0.24.0
-                system = self._ez_client.find_unique_system(
-                    name=require_name,
-                    **kwargs,
+            systems = self._ez_client.find_systems(name=require_name, **kwargs)
+            valid_versions = list(
+                require_version.filter(
+                    [str(Version(system.version)) for system in systems]
                 )
+            )
+            if valid_versions:
+                system_candidates = [
+                    system for system in systems if system.version in valid_versions
+                ]
+                system = system_candidates[0]
+                for system_candidate in system_candidates:
+                    if Version(system_candidate.version) > Version(system.version):
+                        system = system_candidate
         else:
             system = self._ez_client.find_unique_system(
-                name=require_name,
-                **kwargs,
+                name=require_name, filter_latest=True, **kwargs
             )
 
         return system
@@ -493,7 +426,7 @@ class Plugin(object):
         wait_time = 0.1
         while timeout > 0:
             system = self.get_system_matching_version(
-                require, filter_latest=True, filter_running=True, local=True
+                require, filter_running=True, local=True
             )
             if system:
                 self.logger.debug(f"Found system: {system}")